% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sykehus_fig2.R
\name{Data_AndelAntibiotikaTilForebyggingOgBehandling}
\alias{Data_AndelAntibiotikaTilForebyggingOgBehandling}
\title{Data_AndelAntibiotikaTilForebyggingOgBehandling}
\usage{
Data_AndelAntibiotikaTilForebyggingOgBehandling(data, arg)
}
\arguments{
\item{data}{a}

\item{arg}{a}
}
\description{
Data_AndelAntibiotikaTilForebyggingOgBehandling
}
