% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntibiotikaTilXX.R
\name{Data_AntibiotikaTilXX}
\alias{Data_AntibiotikaTilXX}
\title{Data_AntibiotikaTilXX}
\usage{
Data_AntibiotikaTilXX(di, da, DATE_USE, indikasjon = NULL,
  xforebyggingVsBehandling = "Forebygging", ab = "ABBredMethAndre",
  leftVsRightVar = "SykehusKlassifisering")
}
\arguments{
\item{di}{a}

\item{da}{a}

\item{DATE_USE}{a}

\item{indikasjon}{a}

\item{xforebyggingVsBehandling}{a}

\item{ab}{a}

\item{leftVsRightVar}{a}
}
\description{
Data_AntibiotikaTilXX
}
