% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1.R
\name{Data_DeltagelseForekomstHAIogABiSykehjem}
\alias{Data_DeltagelseForekomstHAIogABiSykehjem}
\title{Data_DeltagelseForekomstHAIogABiSykehjem}
\usage{
Data_DeltagelseForekomstHAIogABiSykehjem(data, arg)
}
\arguments{
\item{data}{a}

\item{arg}{sykehjem}
}
\description{
Data_DeltagelseForekomstHAIogABiSykehjem
}
