% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sykehjem_fig1.R
\name{Data_ForekomstHAIiSykehjemPerAvdelingstype}
\alias{Data_ForekomstHAIiSykehjemPerAvdelingstype}
\title{Data_ForekomstHAIiSykehjemPerAvdelingstype}
\usage{
Data_ForekomstHAIiSykehjemPerAvdelingstype(data, arg)
}
\arguments{
\item{data}{a}

\item{arg}{a}
}
\description{
Data_ForekomstHAIiSykehjemPerAvdelingstype
}
