% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForskrivningerAvAntibiotikaPerIndikasjon.R
\name{Data_ForskrivningerAvAntibiotikaPerIndikasjon}
\alias{Data_ForskrivningerAvAntibiotikaPerIndikasjon}
\title{Data_ForskrivningerAvAntibiotikaPerIndikasjon}
\usage{
Data_ForskrivningerAvAntibiotikaPerIndikasjon(di, da, DATE_USE)
}
\arguments{
\item{di}{a}

\item{da}{a}

\item{DATE_USE}{a}
}
\description{
Data_ForskrivningerAvAntibiotikaPerIndikasjon
}
