% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ForskrivningerAvAntibiotikaTilForebyggingOgBehandling.R
\name{Data_ForskrivningerAvAntibiotikaTilForebyggingOgBehandling}
\alias{Data_ForskrivningerAvAntibiotikaTilForebyggingOgBehandling}
\title{Data_ForskrivningerAvAntibiotikaTilForebyggingOgBehandling}
\usage{
Data_ForskrivningerAvAntibiotikaTilForebyggingOgBehandling(di, da, DATE_USE)
}
\arguments{
\item{di}{a}

\item{da}{a}

\item{DATE_USE}{a}
}
\description{
Data_ForskrivningerAvAntibiotikaTilForebyggingOgBehandling
}
