% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sykehus_fig1.R
\name{Data_PrevalensHAIPerSpesialitet}
\alias{Data_PrevalensHAIPerSpesialitet}
\title{Data_PrevalensHAIPerSpesialitet}
\usage{
Data_PrevalensHAIPerSpesialitet(data, arg)
}
\arguments{
\item{data}{a}

\item{arg}{a}
}
\description{
Data_PrevalensHAIPerSpesialitet
}
