% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sykehus_fig2.R
\name{Figure_AndelAntibiotikaTilForebyggingOgBehandling}
\alias{Figure_AndelAntibiotikaTilForebyggingOgBehandling}
\title{Figure_AndelAntibiotikaTilForebyggingOgBehandling}
\usage{
Figure_AndelAntibiotikaTilForebyggingOgBehandling(data, arg)
}
\arguments{
\item{data}{a}

\item{arg}{a}
}
\description{
Figure_AndelAntibiotikaTilForebyggingOgBehandling
}
