% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntibiotikaTilXX.R
\name{Figure_AntibiotikaTilBehandling}
\alias{Figure_AntibiotikaTilBehandling}
\title{Figure_AntibiotikaTilBehandling}
\usage{
Figure_AntibiotikaTilBehandling(data, arg)
}
\arguments{
\item{data}{a}

\item{arg}{a}
}
\description{
Figure_AntibiotikaTilBehandling
}
