% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntibiotikaTilBehandlingOverTid.R
\name{Figure_AntibiotikaTilBehandlingOverTid1}
\alias{Figure_AntibiotikaTilBehandlingOverTid1}
\title{Figure_AntibiotikaTilBehandlingOverTid1}
\usage{
Figure_AntibiotikaTilBehandlingOverTid1(di, da, da_all, DATE_USE)
}
\arguments{
\item{di}{a}

\item{da}{a}

\item{da_all}{a}

\item{DATE_USE}{a}
}
\description{
Figure_AntibiotikaTilBehandlingOverTid1
}
