% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntibiotikaTilBehandlingOverTid.R
\name{Figure_AntibiotikaTilBehandlingOverTid2}
\alias{Figure_AntibiotikaTilBehandlingOverTid2}
\title{Figure_AntibiotikaTilBehandlingOverTid2}
\usage{
Figure_AntibiotikaTilBehandlingOverTid2(di, da, DATE_USE,
  klassifisering = "Samfunnservervet infeksjon")
}
\arguments{
\item{di}{a}

\item{da}{a}

\item{DATE_USE}{a}

\item{klassifisering}{a}
}
\description{
Figure_AntibiotikaTilBehandlingOverTid2
}
