% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntibiotikaTilXX.R
\name{Figure_AntibiotikaTilForebygging}
\alias{Figure_AntibiotikaTilForebygging}
\title{Figure_AntibiotikaTilForebygging}
\usage{
Figure_AntibiotikaTilForebygging(di, da, DATE_USE)
}
\arguments{
\item{di}{a}

\item{da}{a}

\item{DATE_USE}{a}
}
\description{
Figure_AntibiotikaTilForebygging
}
