% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sykehjem_fig1.R
\name{Figure_ForekomstHAIiSykehjemPerAvdelingstype}
\alias{Figure_ForekomstHAIiSykehjemPerAvdelingstype}
\title{Figure_ForekomstHAIiSykehjemPerAvdelingstype}
\usage{
Figure_ForekomstHAIiSykehjemPerAvdelingstype(data, arg)
}
\arguments{
\item{data}{a}

\item{arg}{a}
}
\description{
Figure_ForekomstHAIiSykehjemPerAvdelingstype
}
