% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForskrivningerAvAntibiotikaPerIndikasjon.R
\name{Figure_ForskrivningerAvAntibiotikaPerIndikasjon}
\alias{Figure_ForskrivningerAvAntibiotikaPerIndikasjon}
\title{Figure_ForskrivningerAvAntibiotikaPerIndikasjon}
\usage{
Figure_ForskrivningerAvAntibiotikaPerIndikasjon(di, da, DATE_USE,
  indikasjon = "Nedre urinveisinfeksjon", display_name = indikasjon,
  with_metenamin = T)
}
\arguments{
\item{di}{a}

\item{da}{a}

\item{DATE_USE}{a}

\item{indikasjon}{a}

\item{display_name}{a}

\item{with_metenamin}{a}
}
\description{
Figure_ForskrivningerAvAntibiotikaPerIndikasjon
}
