% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sykehus_fig3.R
\name{Figure_ForskrivningerAvAntibiotikaTilBehandlingPerIndikasjon}
\alias{Figure_ForskrivningerAvAntibiotikaTilBehandlingPerIndikasjon}
\title{Figure_ForskrivningerAvAntibiotikaTilBehandlingPerIndikasjon}
\usage{
Figure_ForskrivningerAvAntibiotikaTilBehandlingPerIndikasjon(data, arg)
}
\arguments{
\item{data}{a}

\item{arg}{a}
}
\description{
Figure_ForskrivningerAvAntibiotikaTilBehandlingPerIndikasjon
}
