% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sykehus_fig3.R
\name{Figure_ForskrivningerAvAntibiotikaTilBehandlingPerSpesialitet}
\alias{Figure_ForskrivningerAvAntibiotikaTilBehandlingPerSpesialitet}
\title{Figure_ForskrivningerAvAntibiotikaTilBehandlingPerSpesialitet}
\usage{
Figure_ForskrivningerAvAntibiotikaTilBehandlingPerSpesialitet(data, arg)
}
\arguments{
\item{data}{a}

\item{arg}{a}
}
\description{
Figure_ForskrivningerAvAntibiotikaTilBehandlingPerSpesialitet
}
