% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ForskrivningerAvAntibiotikaTilForebyggingOgBehandling.R
\name{Figure_ForskrivningerAvAntibiotikaTilForebyggingOgBehandling}
\alias{Figure_ForskrivningerAvAntibiotikaTilForebyggingOgBehandling}
\title{Figure_ForskrivningerAvAntibiotikaTilForebyggingOgBehandling}
\usage{
Figure_ForskrivningerAvAntibiotikaTilForebyggingOgBehandling(di, da, DATE_USE)
}
\arguments{
\item{di}{a}

\item{da}{a}

\item{DATE_USE}{a}
}
\description{
Figure_ForskrivningerAvAntibiotikaTilForebyggingOgBehandling
}
