% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sykehjem_fig2.R
\name{Figure_ForskrivningerAvAntibiotikaTilForebyggingOgBehandlingPerIndikasjon}
\alias{Figure_ForskrivningerAvAntibiotikaTilForebyggingOgBehandlingPerIndikasjon}
\title{Figure_ForskrivningerAvAntibiotikaTilForebyggingOgBehandlingPerIndikasjon}
\usage{

  Figure_ForskrivningerAvAntibiotikaTilForebyggingOgBehandlingPerIndikasjon(data,
  arg)
}
\arguments{
\item{data}{a}

\item{arg}{a}
}
\description{
Figure_ForskrivningerAvAntibiotikaTilForebyggingOgBehandlingPerIndikasjon
}
