% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sykehus_fig1.R
\name{Figure_PrevalensHAIPerSpesialitet}
\alias{Figure_PrevalensHAIPerSpesialitet}
\title{Figure_PrevalensHAIPerSpesialitet}
\usage{
Figure_PrevalensHAIPerSpesialitet(data, arg)
}
\arguments{
\item{data}{a}

\item{arg}{a}
}
\description{
Figure_PrevalensHAIPerSpesialitet
}
