% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{gen_plan_email}
\alias{gen_plan_email}
\title{gen_plan_email}
\usage{
gen_plan_email(dev = TRUE, abonnenter, DATE_USE)
}
\arguments{
\item{dev}{a}

\item{abonnenter}{a}

\item{DATE_USE}{a}
}
\description{
gen_plan_email
}
