% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{gen_stack_sykehjem}
\alias{gen_stack_sykehjem}
\title{GenStackSykehjem}
\usage{
gen_stack_sykehjem(
  dev = FALSE,
  outputDir = fd::path("results"),
  FILES_RMD_USE_SYKEHJEM,
  FILES_RMD_USE_SYKEHUS,
  requested_date = NULL,
  abonnenter_file
)
}
\arguments{
\item{dev}{a}

\item{outputDir}{a}

\item{FILES_RMD_USE_SYKEHJEM}{a}

\item{FILES_RMD_USE_SYKEHUS}{a}

\item{requested_date}{a}

\item{abonnenter_file}{a}
}
\description{
GenStackSykehjem
}
