% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{gen_stack_sykehus}
\alias{gen_stack_sykehus}
\title{GenStackSykehus}
\usage{
gen_stack_sykehus(dev = FALSE, outputDir = fd::path("results"),
  FILES_RMD_USE_SYKEHJEM, FILES_RMD_USE_SYKEHUS, requested_date = NULL)
}
\arguments{
\item{dev}{a}

\item{outputDir}{a}

\item{FILES_RMD_USE_SYKEHJEM}{a}

\item{FILES_RMD_USE_SYKEHUS}{a}

\item{requested_date}{a}
}
\description{
GenStackSykehus
}
