% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Render.R
\name{GenStackSykehus}
\alias{GenStackSykehus}
\title{GenStackSykehus}
\usage{
GenStackSykehus(dev = FALSE, outputDir = fhi::DashboardFolder("results"),
  FILES_RMD_USE_SYKEHJEM, FILES_RMD_USE_SYKEHUS)
}
\arguments{
\item{dev}{a}

\item{outputDir}{a}

\item{FILES_RMD_USE_SYKEHJEM}{a}

\item{FILES_RMD_USE_SYKEHUS}{a}
}
\description{
GenStackSykehus
}
