% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaGlobalVariables.R
\name{SetConfig}
\alias{SetConfig}
\title{If folders are setup according to the
dashboard philosophy, then this function
sets RPROJ}
\usage{
SetConfig(var, val)
}
\arguments{
\item{var}{a}

\item{val}{a}
}
\description{
If folders are setup according to the
dashboard philosophy, then this function
sets RPROJ
}
