% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1.R
\name{Table_DeltagelseForekomstHAIogABiSykehjem}
\alias{Table_DeltagelseForekomstHAIogABiSykehjem}
\title{Table_DeltagelseForekomstHAIogABiSykehjem}
\usage{
Table_DeltagelseForekomstHAIogABiSykehjem(data, arg)
}
\arguments{
\item{data}{a}

\item{arg}{a}
}
\description{
Table_DeltagelseForekomstHAIogABiSykehjem
}
