% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Sykehus_AndelAntibiotikaTilForebyggingOgBehandling.R
\name{Data_AndelAntibiotikaTilForebyggingOgBehandling}
\alias{Data_AndelAntibiotikaTilForebyggingOgBehandling}
\title{Data_AndelAntibiotikaTilForebyggingOgBehandling}
\usage{
Data_AndelAntibiotikaTilForebyggingOgBehandling(di, da, DATE_USE)
}
\arguments{
\item{di}{a}

\item{da}{a}

\item{DATE_USE}{a}
}
\description{
Data_AndelAntibiotikaTilForebyggingOgBehandling
}
