% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntibiotikaTilBehandlingOverTid.R
\name{Data_AntibiotikaTilBehandlingOverTid}
\alias{Data_AntibiotikaTilBehandlingOverTid}
\title{Data_AntibiotikaTilBehandlingOverTid}
\usage{
Data_AntibiotikaTilBehandlingOverTid(di, da, indikasjon = NULL,
  ab = "sykehusAB2", klassifisering = NULL)
}
\arguments{
\item{di}{a}

\item{da}{a}

\item{indikasjon}{a}

\item{ab}{a}

\item{klassifisering}{a}

\item{DATE_USE}{a}
}
\description{
Data_AntibiotikaTilBehandlingOverTid
}
