% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltagelseForekomstHAIogABiSykehjem.R
\name{Data_DeltagelseForekomstHAIogABiSykehjem}
\alias{Data_DeltagelseForekomstHAIogABiSykehjem}
\title{Data_DeltagelseForekomstHAIogABiSykehjem}
\usage{
Data_DeltagelseForekomstHAIogABiSykehjem(di, da, level, DATE_USE)
}
\arguments{
\item{di}{a}

\item{da}{a}

\item{level}{a}

\item{DATE_USE}{a}
}
\description{
Data_DeltagelseForekomstHAIogABiSykehjem
}
