% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForekomstHAIiSykehjemPerAvdelingstype.R
\name{Data_ForekomstHAIiSykehjemPerAvdelingstype}
\alias{Data_ForekomstHAIiSykehjemPerAvdelingstype}
\title{Data_ForekomstHAIiSykehjemPerAvdelingstype}
\usage{
Data_ForekomstHAIiSykehjemPerAvdelingstype(di, da, DATE_USE)
}
\arguments{
\item{di}{a}

\item{da}{a}

\item{DATE_USE}{a}
}
\description{
Data_ForekomstHAIiSykehjemPerAvdelingstype
}
