% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sykehjem_fig3.R
\name{Data_ForskrivningerAvAntibiotikaPerIndikasjon}
\alias{Data_ForskrivningerAvAntibiotikaPerIndikasjon}
\title{Data_ForskrivningerAvAntibiotikaPerIndikasjon}
\usage{
Data_ForskrivningerAvAntibiotikaPerIndikasjon(data, arg)
}
\arguments{
\item{data}{a}

\item{arg}{a}
}
\description{
Data_ForskrivningerAvAntibiotikaPerIndikasjon
}
