% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sykehus_fig3.R
\name{Data_ForskrivningerAvAntibiotikaTilForebyggingOgBehandling}
\alias{Data_ForskrivningerAvAntibiotikaTilForebyggingOgBehandling}
\title{Data_ForskrivningerAvAntibiotikaTilForebyggingOgBehandling}
\usage{
Data_ForskrivningerAvAntibiotikaTilForebyggingOgBehandling(data, arg)
}
\arguments{
\item{data}{a}

\item{arg}{a}
}
\description{
Data_ForskrivningerAvAntibiotikaTilForebyggingOgBehandling
}
