% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrevalensHAIPerSpesialitet.R
\name{Data_PrevalensHAIPerSpesialitet}
\alias{Data_PrevalensHAIPerSpesialitet}
\title{Data_PrevalensHAIPerSpesialitet}
\usage{
Data_PrevalensHAIPerSpesialitet(di, da, level, DATE_USE)
}
\arguments{
\item{di}{a}

\item{da}{a}

\item{level}{a}

\item{DATE_USE}{a}
}
\description{
Data_PrevalensHAIPerSpesialitet
}
