% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Sykehus_AndelAntibiotikaTilForebyggingOgBehandling.R
\name{Figure_AndelAntibiotikaTilForebyggingOgBehandling}
\alias{Figure_AndelAntibiotikaTilForebyggingOgBehandling}
\title{Figure_AndelAntibiotikaTilForebyggingOgBehandling}
\usage{
Figure_AndelAntibiotikaTilForebyggingOgBehandling(di, da, DATE_USE)
}
\arguments{
\item{di}{a}

\item{da}{a}

\item{DATE_USE}{a}
}
\description{
Figure_AndelAntibiotikaTilForebyggingOgBehandling
}
