% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntibiotikaTilXX.R
\name{Figure_AntibiotikaTilBehandling}
\alias{Figure_AntibiotikaTilBehandling}
\title{Figure_AntibiotikaTilBehandling}
\usage{
Figure_AntibiotikaTilBehandling(di, da, DATE_USE, indikasjon = NULL,
  ab = "ABBredMethAndre", leftVsRightVar = "SykehusKlassifisering")
}
\arguments{
\item{di}{a}

\item{da}{a}

\item{DATE_USE}{a}

\item{indikasjon}{a}

\item{ab}{a}

\item{leftVsRightVar}{a}
}
\description{
Figure_AntibiotikaTilBehandling
}
