% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sykehus_fig5.R
\name{Figure_AntibiotikaTilBehandlingOverTid1}
\alias{Figure_AntibiotikaTilBehandlingOverTid1}
\title{Figure_AntibiotikaTilBehandlingOverTid1}
\usage{
Figure_AntibiotikaTilBehandlingOverTid1(data, arg)
}
\arguments{
\item{data}{a}

\item{arg}{a}
}
\description{
Figure_AntibiotikaTilBehandlingOverTid1
}
