% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sykehus_fig5.R
\name{Figure_AntibiotikaTilBehandlingOverTid2}
\alias{Figure_AntibiotikaTilBehandlingOverTid2}
\title{Figure_AntibiotikaTilBehandlingOverTid2}
\usage{
Figure_AntibiotikaTilBehandlingOverTid2(data, arg)
}
\arguments{
\item{data}{a}

\item{arg}{a}
}
\description{
Figure_AntibiotikaTilBehandlingOverTid2
}
