% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForekomstHAIiSykehjemPerAvdelingstype.R
\name{Figure_ForekomstHAIiSykehjemPerAvdelingstype}
\alias{Figure_ForekomstHAIiSykehjemPerAvdelingstype}
\title{Figure_ForekomstHAIiSykehjemPerAvdelingstype}
\usage{
Figure_ForekomstHAIiSykehjemPerAvdelingstype(di, da, DATE_USE)
}
\arguments{
\item{di}{a}

\item{da}{a}

\item{DATE_USE}{a}
}
\description{
Figure_ForekomstHAIiSykehjemPerAvdelingstype
}
