% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sykehjem_fig3.R
\name{Figure_ForskrivningerAvAntibiotikaPerIndikasjon}
\alias{Figure_ForskrivningerAvAntibiotikaPerIndikasjon}
\title{Figure_ForskrivningerAvAntibiotikaPerIndikasjon}
\usage{
Figure_ForskrivningerAvAntibiotikaPerIndikasjon(data, arg)
}
\arguments{
\item{data}{a}

\item{arg}{a}
}
\description{
Figure_ForskrivningerAvAntibiotikaPerIndikasjon
}
