% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ForskrivningerAvAntibiotikaTilForebyggingOgBehandling.R
\name{Figure_ForskrivningerAvAntibiotikaTilBehandlingPerSpesialitet}
\alias{Figure_ForskrivningerAvAntibiotikaTilBehandlingPerSpesialitet}
\title{Figure_ForskrivningerAvAntibiotikaTilBehandlingPerSpesialitet}
\usage{
Figure_ForskrivningerAvAntibiotikaTilBehandlingPerSpesialitet(di, da,
  DATE_USE)
}
\arguments{
\item{di}{a}

\item{da}{a}

\item{DATE_USE}{a}
}
\description{
Figure_ForskrivningerAvAntibiotikaTilBehandlingPerSpesialitet
}
