% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/ForskrivningerAvAntibiotikaTilForebyggingOgBehandling.R
\name{Figure_ForskrivningerAvAntibiotikaTilForebyggingOgBehandlingPerIndikasjon}
\alias{Figure_ForskrivningerAvAntibiotikaTilForebyggingOgBehandlingPerIndikasjon}
\title{Figure_ForskrivningerAvAntibiotikaTilForebyggingOgBehandlingPerIndikasjon}
\usage{
Figure_ForskrivningerAvAntibiotikaTilForebyggingOgBehandlingPerIndikasjon(di,
  da, DATE_USE)
}
\arguments{
\item{di}{a}

\item{da}{a}

\item{DATE_USE}{a}
}
\description{
Figure_ForskrivningerAvAntibiotikaTilForebyggingOgBehandlingPerIndikasjon
}
