% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrevalensHAIPerSpesialitet.R
\name{Figure_PrevalensHAIPerSpesialitet}
\alias{Figure_PrevalensHAIPerSpesialitet}
\title{Figure_PrevalensHAIPerSpesialitet}
\usage{
Figure_PrevalensHAIPerSpesialitet(di, da, level, DATE_USE)
}
\arguments{
\item{di}{a}

\item{da}{a}

\item{level}{a}

\item{DATE_USE}{a}
}
\description{
Figure_PrevalensHAIPerSpesialitet
}
