% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Render.R
\name{GenStackSykehjem}
\alias{GenStackSykehjem}
\title{GenStackSykehjem}
\usage{
GenStackSykehjem(dev = FALSE, outputDir = fhi::DashboardFolder("results"),
  FILES_RMD_USE_SYKEHJEM, FILES_RMD_USE_SYKEHUS)
}
\arguments{
\item{dev}{a}

\item{outputDir}{a}

\item{FILES_RMD_USE_SYKEHJEM}{a}

\item{FILES_RMD_USE_SYKEHUS}{a}
}
\description{
GenStackSykehjem
}
