% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltagelseForekomstHAIogABiSykehjem.R,
%   R/PrevalensHAIPerSpesialitet.R
\name{Table_DeltagelseForekomstHAIogABiSykehjem}
\alias{Table_DeltagelseForekomstHAIogABiSykehjem}
\alias{Table_DeltagelseForekomstHAIogABiSykehjem}
\title{Table_DeltagelseForekomstHAIogABiSykehjem}
\usage{
Table_DeltagelseForekomstHAIogABiSykehjem(di, da, level, DATE_USE,
  type = "sykehjem")

Table_DeltagelseForekomstHAIogABiSykehjem(di, da, level, DATE_USE,
  type = "sykehjem")
}
\arguments{
\item{di}{a}

\item{da}{a}

\item{level}{a}

\item{DATE_USE}{a}

\item{type}{sykehjem}

\item{di}{a}

\item{da}{a}

\item{level}{a}

\item{DATE_USE}{a}

\item{type}{sykehjem}
}
\description{
Table_DeltagelseForekomstHAIogABiSykehjem

Table_DeltagelseForekomstHAIogABiSykehjem
}
